pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aoc_runner"
version = "0.0.1"
authors = [
  { name="Brandon Hippe", email="bphippe@gmail.com" },
]
description = "Runner for Advent of Code in multiple languages, with capabilities for viewing solutions and runtimes."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
  "advent-of-code-ocr==1.0.0",
  "attrs==25.1.0",
  "beautifulsoup4==4.13.3",
  "certifi==2025.1.31",
  "charset-normalizer==3.4.1",
  "contourpy==1.3.1",
  "cycler==0.12.1",
  "fonttools==4.56.0",
  "idna==3.10",
  "kiwisolver==1.4.8",
  "matplotlib==3.10.0",
  "numpy==2.2.2",
  "opencv-python==4.11.0.86",
  "packaging==24.2",
  "pillow==11.1.0",
  "prettytable==3.14.0",
  "pyarrow==19.0.0",
  "pyparsing==3.2.1",
  "pytesseract==0.3.13",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.0.1",
  "PyYAML==6.0.2",
  "requests==2.32.3",
  "rerun-sdk==0.22.0",
  "six==1.17.0",
  "soupsieve==2.6",
  "typing_extensions==4.12.2",
  "urllib3==2.3.0",
  "wcwidth==0.2.13",
]

[project.scripts]
aoc = "runner:main"

# [project.urls]
# Homepage = "https://github.com/pypa/sampleproject"
# Issues = "https://github.com/pypa/sampleproject/issues"